+++ CHANGELOG.md
@@ -9,6 +9,10 @@ All notable changes to this project will be documented in this file.
 - `minimize-distance` and `minimize-duration` objectives
 - `init-size` option to control amount of initial solutions to be built built from CLI
 
+### Changed
+
+- optimize cluster ruin method
+
 
 ## [v1.9.1] - 2021-03-24